enable working in specified dir

+++ analysis/workflow/scripts/aou/get_pgen.bash
@@ -5,6 +5,7 @@ set -euo pipefail
 region="$1"
 out_prefix="$2"
 threads="${3:-1}"
+TEMP_DIR="$4" # default: tmp dir
 
 export GCS_REQUESTER_PAYS_PROJECT="${GOOGLE_PROJECT}"
 export GCS_OAUTH_TOKEN="$(gcloud auth application-default print-access-token)"
@@ -16,20 +17,25 @@ end="$(echo "$region" | cut -f2 -d: | cut -f2 -d-)"
 region=chr"${chrom}:${pos}-${end}"
 VCF_DIR="${WORKSPACE_BUCKET}/beagle_hg38/chr${chrom}/chr${chrom}."'BATCH*_output.vcf.gz'
 CDR_DIR="gs://fc-aou-datasets-controlled/v7"
-TEMP_DIR=$(mktemp -d)
-trap 'rm -rf "$TEMP_DIR"' EXIT
+if [[ -n "$TEMP_DIR" ]]; then
+    mkdir -p "$TEMP_DIR"
+else
+    TEMP_DIR="$(mktemp -d)"
+    trap 'rm -rf "$TEMP_DIR"' EXIT
+fi
+mkdir -p "$TEMP_DIR/$region"
 
 batches="$(gsutil ls "$VCF_DIR" | grep -oP '(?<=BATCH)\d+' | sort -n)"
 cd "$TEMP_DIR"
-echo "tempdir: $TEMP_DIR"
-echo "$batches" | xargs -P <num_parallel_jobs> -I{} sh -c '
+echo "workdir: $TEMP_DIR"
+echo "$batches" | xargs -P "$threads" -I{} sh -c '
   batch="{}"
-  bcftools view --threads 1 -O b -o "$batch.bcf" -r "$region" "$(echo "$VCF_DIR" | sed '\''s/*/'"$batch"'/\'')"
+  bcftools view --threads 1 -O b -o "$region/batch.bcf" -r "$region" "$(echo "$VCF_DIR" | sed '\''s/*/'"$batch"'/\'')"
 '
 
 # assert that all batches were downloaded
 for batch in $batches; do
-    file_path="$batch".bcf
+    file_path="$region/$batch".bcf
     if [ ! -f "$file_path" ]; then
         echo "Error: Required file not found at $file_path" >&2
         exit 1
@@ -37,5 +43,5 @@ for batch in $batches; do
 done
 
 cd -
-bcftools merge --threads "$threads" --no-index -O b -o "$out_prefix".bcf -l <(ls "$TEMP_DIR"/*.bcf)
+bcftools merge --threads "$threads" --no-index -O b -o "$out_prefix".bcf -l <(ls "$TEMP_DIR/$region"/*.bcf)
 plink2 --threads "$threads" --out "$out_prefix" --nonfounders --bcf "$out_prefix".bcf --geno 0 --make-pgen --allow-extra-chr --max-alleles 2 --chr "$chrom" --from-bp "$pos" --to-bp "$end"
